+++ indico/modules/rb/views/__init__.py
@@ -26,6 +26,7 @@ from indico.web.views import WPDecorated, WPJinjaMixin
 
 class WPRoomBookingBase(WPJinjaMixin, WPDecorated):
     template_prefix = 'rb/'
+    title = _('Room Booking')
 
     def __init__(self, rh, **kwargs):
         kwargs['active_menu_item'] = self.sidemenu_option
@@ -36,9 +37,6 @@ class WPRoomBookingBase(WPJinjaMixin, WPDecorated):
 
     def _get_breadcrumbs(self):
         return render_breadcrumbs(_('Room Booking'))
-
-    def _getTitle(self):
-        return '{} - {}'.format(WPDecorated._getTitle(self), _('Room Booking'))
 
     def _getBody(self, params):
         return self._getPageContent(params)